Enable key sizes larger than 127 bytes in the file format
  Modify index, record, and tombstone file serialization so that the
  prior version byte and key size byte now represent a 5 bit version
  and an 11 bit key length.

  Because we are at version zero,
  old code can still read data written by new code as long as
  key sizes are not larger than 127 bytes.  If so, old code
  will interpret large key sizes as an incompatible version number.

  Data written by old code is fully compatible with new code because
  a version 0 byte still represents 0 after this change, and also does
  not impact the key.

  Prior:
  version byte -> version number, 0 to 255
  key size byte -> key size, 0 to 127

  After this commit:
  version byte -> top 5 bits are the version number, 0 to 31
                  bottom 3 bits are the three MSB of key size
  key size byte -> lower 8 bits of key size

  -------------------
  |  vbyte | ksbyte |   old version and key size -- 8 bits each
  -------------------
  | ver | key size  |   new version and key size -- 5 / 11 bit split
  -------------------

+++ src/main/java/com/oath/halodb/Versions.java
@@ -7,8 +7,8 @@ package com.oath.halodb;
 
 class Versions {
 
-    static final int CURRENT_DATA_FILE_VERSION = 0;
-    static final int CURRENT_INDEX_FILE_VERSION = 0;
-    static final int CURRENT_TOMBSTONE_FILE_VERSION = 0;
-    static final int CURRENT_META_FILE_VERSION = 0;
+    static final byte CURRENT_DATA_FILE_VERSION = 0;
+    static final byte CURRENT_INDEX_FILE_VERSION = 0;
+    static final byte CURRENT_TOMBSTONE_FILE_VERSION = 0;
+    static final byte CURRENT_META_FILE_VERSION = 0;
 }
